pyproject.toml
[tool.poetry]
name = "socialfeya"
version = "0.1.0"
description = "Social Network by AleRyp"
authors = ["Andrian Kozynets"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.3"
djangorestframework = "^3.12.1"
Pillow = "^8.0.1"
psycopg2 = "^2.8.6"
djoser = "^2.1.0"
djangorestframework-simplejwt = "^4.4.0"
drf-yasg = "^1.20.0"
django-cors-headers = "^3.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
